[example-app] update to Next 14, refactor pages, site-wide search
