+++ package.json
@@ -11,19 +11,19 @@
     "packages/*"
   ],
   "dependencies": {
-    "@rollup/plugin-terser": "^0.4.3",
-    "@rollup/plugin-typescript": "^11.0.0",
-    "@types/react": "^18.2.7",
+    "@rollup/plugin-terser": "^0.4.4",
+    "@rollup/plugin-typescript": "^11.1.5",
+    "@types/react": "^18.2.37",
     "eslint": "^8.53.0",
     "eslint-config-universe": "^12.0.0",
     "lerna": "^7.4.2",
     "npm-run-all": "^4.1.5",
     "prettier": "^3.0.3",
     "react": "^18.2.0",
-    "rimraf": "^5.0.0",
+    "rimraf": "^5.0.5",
     "rollup": "^3.23.0",
     "rollup-plugin-copy": "^3.4.0",
-    "typescript": "^4.9.5"
+    "typescript": "^5.2.2"
   },
   "eslintConfig": {
     "root": true,
