+++ packages/example-web/components/SidebarLink.tsx
@@ -5,16 +5,19 @@ import { useRouter } from 'next/router';
 type Props = {
   href: string;
   text: string;
+  size?: 'sm' | 'md';
 };
 
-export function SidebarLink({ href, text }: Props) {
+export function SidebarLink({ href, text, size = 'md' }: Props) {
   const { pathname } = useRouter();
   return (
     <Link
       href={href}
       className={mergeClasses(
-        'heading-xl font-medium flex items-center gap-2 transition-colors hover:underline',
-        pathname === href && 'text-link'
+        'font-medium flex items-center gap-2 transition-colors hover:underline',
+        pathname === href && 'text-link',
+        size === 'sm' && 'heading-base pl-2',
+        size === 'md' && 'heading-xl'
       )}>
       <span>{text}</span>
     </Link>
