+++ packages/example-web/data/quickActionEntries.ts
@@ -1,0 +1,11 @@
+import { ThemeIcon } from '@expo/styleguide-icons';
+
+import { QuickActionToggleThemeItem } from '@/components/search/QuickActionItem';
+
+export const entries = [
+  {
+    label: 'Toggle Theme',
+    Icon: ThemeIcon,
+    Element: QuickActionToggleThemeItem,
+  },
+];
