+++ packages/example-web/data/styleguideEntries.ts
@@ -1,0 +1,45 @@
+import { ExpoGoLogo } from '@expo/styleguide';
+import { BezierCurve03Icon, LayoutAlt03Icon, PaletteIcon, TextInputIcon, Type01Icon } from '@expo/styleguide-icons';
+
+export const entries = [
+  {
+    label: 'Colors: Palette',
+    url: '/colors#palette',
+    Icon: PaletteIcon,
+  },
+  {
+    label: 'Colors: Project Background',
+    url: '/colors#project',
+    Icon: PaletteIcon,
+  },
+  {
+    label: 'Typography: Headings',
+    url: '/typography#headings',
+    Icon: Type01Icon,
+  },
+  {
+    label: 'Typography: Elements',
+    url: '/typography#elements',
+    Icon: Type01Icon,
+  },
+  {
+    label: 'Icons: Logos',
+    url: '/icons#logos',
+    Icon: ExpoGoLogo,
+  },
+  {
+    label: 'Icons: Icon Set',
+    url: '/icons#set',
+    Icon: BezierCurve03Icon,
+  },
+  {
+    label: 'UI: Components',
+    url: '/ui/components',
+    Icon: LayoutAlt03Icon,
+  },
+  {
+    label: 'UI: Search',
+    url: '/ui/search',
+    Icon: TextInputIcon,
+  },
+];
