+++ packages/example-web/next.config.js
@@ -5,8 +5,14 @@ const withBundleAnalyzer = require('@next/bundle-analyzer')({
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-  experimental: {
-    fontLoaders: [{ loader: '@next/font/google', options: { subsets: ['latin'] } }],
+  async redirects() {
+    return [
+      {
+        source: '/ui',
+        destination: '/ui/components',
+        permanent: true,
+      },
+    ];
   },
 };
