+++ packages/example-web/package.json
@@ -9,21 +9,20 @@
     "clean": "rimraf .next"
   },
   "dependencies": {
-    "@next/font": "13.1.6",
-    "@types/node": "18.13.0",
-    "@types/react": "18.0.28",
-    "@types/react-dom": "18.0.10",
-    "next": "13.1.6",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "@types/node": "^18.18.9",
+    "@types/react": "^18.2.37",
+    "@types/react-dom": "^18.2.15",
+    "next": "^14.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "tailwind-merge": "^1.13.2",
-    "typescript": "4.9.5"
+    "typescript": "^5.2.2"
   },
   "devDependencies": {
-    "@next/bundle-analyzer": "13.1.6",
-    "autoprefixer": "^10.4.13",
-    "eslint-config-next": "^13.1.6",
-    "postcss": "^8.4.21",
+    "@next/bundle-analyzer": "^14.0.2",
+    "autoprefixer": "^10.4.16",
+    "eslint-config-next": "^14.0.2",
+    "postcss": "^8.4.31",
     "tailwindcss": "^3.3.2"
   },
   "eslintConfig": {
