+++ packages/example-web/public/global.css
